Fix test error

+++ features/sharding/core/src/test/java/org/apache/shardingsphere/sharding/metadata/data/ShardingTableStatisticsCollectorTest.java
@@ -87,12 +87,12 @@ class ShardingTableStatisticsCollectorTest {
         Collection<Map<String, Object>> actualRows = statisticsCollector.collect("foo_db", "shardingsphere", "sharding_table_statistics", metaData);
         assertFalse(actualRows.isEmpty());
         Collection<Map<String, Object>> expectedRows = new LinkedList<>();
-        expectedRows.add(createRowColumnValues(1, "foo_db", "foo_tbl", "ds_0", "foo_tbl", new BigDecimal("0"), new BigDecimal("0")));
-        expectedRows.add(createRowColumnValues(2, "foo_db", "foo_tbl", "ds_1", "foo_tbl", new BigDecimal("0"), new BigDecimal("0")));
+        expectedRows.add(createRowColumnValues(1L, "foo_db", "foo_tbl", "ds_0", "foo_tbl", new BigDecimal("0"), new BigDecimal("0")));
+        expectedRows.add(createRowColumnValues(2L, "foo_db", "foo_tbl", "ds_1", "foo_tbl", new BigDecimal("0"), new BigDecimal("0")));
         assertRowsValue(expectedRows, actualRows);
     }
     
-    private Map<String, Object> createRowColumnValues(final int id, final String logicDatabaseName, final String logicTableName, final String actualDatabaseName,
+    private Map<String, Object> createRowColumnValues(final long id, final String logicDatabaseName, final String logicTableName, final String actualDatabaseName,
                                                       final String actualTableName, final BigDecimal rowCount, final BigDecimal size) {
         Map<String, Object> result = new HashMap<>();
         result.put("id", id);
